+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/CoreExtensions.kt
@@ -9,6 +9,7 @@ import kotlin.contracts.*
 
 private val logger = KotlinLogging.logger {}
 
+//REGION Global extensions
 
 /**转化为指定类型，或者抛出异常。用于链式调用。*/
 inline fun <reified R> Any?.cast(): R = this as R
@@ -16,54 +17,7 @@ inline fun <reified R> Any?.cast(): R = this as R
 /**转化为指定类型，或者返回null。用于链式调用。*/
 inline fun <reified R> Any?.castOrNull(): R? = this as? R
 
-
-/**从二元素元组构造三元素元组。*/
-infix fun <A, B, C> Pair<A, B>.andTo(that: C): Triple<A, B, C> = Triple(this.first, this.second, that)
-
-/**取在指定范围内的夹值。*/
-infix fun <T : Comparable<T>> T.clamp(range: ClosedRange<T>): T = this.coerceIn(range)
-
-
-@PublishedApi internal var enableOnce = false
-
-/**执行且仅执行一次操作。可指定是否重置单次状态。*/
-@ExperimentalContracts
-inline fun once(resetStatus: Boolean = false, block: () -> Unit) {
-	contract {
-		callsInPlace(block, InvocationKind.EXACTLY_ONCE)
-	}
-	if(resetStatus) enableOnce = false
-	if(enableOnce) return
-	enableOnce = true
-	block()
-}
-
-
-/**如果判定失败，则抛出一个[UnsupportedOperationException]。*/
-@OutlookImplementationApi
-@ExperimentalContracts
-inline fun reject(value: Boolean) {
-	contract {
-		returns() implies value
-	}
-	reject(value) { "Unsupported operation." }
-}
-
-/**如果判定失败，则抛出一个[UnsupportedOperationException]，带有懒加载的信息。*/
-@OutlookImplementationApi
-@ExperimentalContracts
-inline fun reject(value: Boolean, lazyMessage: () -> Any) {
-	contract {
-		returns() implies value
-	}
-	if(!value) {
-		val message = lazyMessage()
-		throw UnsupportedOperationException(message.toString())
-	}
-}
-
-
-//TODO 让TODO插件能够正确地提供高亮
+//REGION Standard.kt extensions
 
 /**表明一个方法体推迟了实现。*/
 @OutlookImplementationApi
@@ -87,6 +41,46 @@ inline fun FIXME() = run { nearestLogger().warn("There is an issue in this opera
 /**打印一段消息，以表明一个方法体中存在问题，并指明原因。*/
 @OutlookImplementationApi
 inline fun FIXME(message: String) = run { nearestLogger().warn("There is an issue in this operation: $message") }
+
+
+@PublishedApi internal var enableOnce = false
+
+/**执行且仅执行一次操作。可指定是否重置单次状态。*/
+@ExperimentalContracts
+inline fun once(resetStatus: Boolean = false, block: () -> Unit) {
+	contract {
+		callsInPlace(block, InvocationKind.EXACTLY_ONCE)
+	}
+	if(resetStatus) enableOnce = false
+	if(enableOnce) return
+	enableOnce = true
+	block()
+}
+
+//REGION Precondition.kt extensions
+
+/**如果判定失败，则抛出一个[UnsupportedOperationException]。*/
+@OutlookImplementationApi
+@ExperimentalContracts
+inline fun reject(value: Boolean) {
+	contract {
+		returns() implies value
+	}
+	reject(value) { "Unsupported operation." }
+}
+
+/**如果判定失败，则抛出一个[UnsupportedOperationException]，带有懒加载的信息。*/
+@OutlookImplementationApi
+@ExperimentalContracts
+inline fun reject(value: Boolean, lazyMessage: () -> Any) {
+	contract {
+		returns() implies value
+	}
+	if(!value) {
+		val message = lazyMessage()
+		throw UnsupportedOperationException(message.toString())
+	}
+}
 
 
 /**得到最近的堆栈追踪信息。即，得到最近一个内联方法的调用处的信息。*/