+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/interfaces/TagEnum.kt
@@ -1,0 +1,6 @@
+package com.windea.breezeframework.core.interfaces
+
+import java.io.*
+
+/**标签枚举的接口。一般的，一个模型类可以拥有多个标签。*/
+interface TagEnum : Serializable