+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/dsl/Dsl.kt
@@ -2,15 +2,23 @@
 
 package com.windea.breezeframework.data.dsl
 
-/////////////Top interfaces
+//规定：
+//所有的Dsl元素和Dsl构建方法都必须添加对应的slMarker注解
+//所有的Dsl构建方法都要尽可能地写成内联形式和X表达式形式，且不要显式声明返回值，使用`Xxx.also{}`的写法
+//所有的Dsl元素的属性都不应该声明在构造方法之内，因为可能需要对输入参数进行处理。
 
-/**Dsl。即，领域专用语言。*/
-interface Dsl
+//REGION Top interfaces
+
+@DslMarker
+internal annotation class GeneralDsl
+
+/**Dsl的构建器。*/
+interface DslBuilder
 
 /**Dsl的配置。*/
 interface DslConfig
 
-//////////////General Dsl interfaces and related extensions
+//REGION General Dsl interfaces and related extensions
 
 /**可换行内容。*/
 interface CanWrapContent {
@@ -22,20 +30,28 @@ interface CanIndentContent {
 	var indentContent: Boolean
 }
 
+
 /**设置是否换行内容。*/
-inline infix fun <T> T.wrap(value: Boolean) where T : CanWrapContent = this.also { wrapContent = value }
-
-/**换行内容。*/
-inline fun <T> T.wrap() where T : CanWrapContent = this.also { wrapContent = true }
-
-/**不换行内容。*/
-inline fun <T> T.unwrap() where T : CanWrapContent = this.also { wrapContent = false }
+@GeneralDsl
+inline infix fun <T> T.wrap(value: Boolean): T where T : CanWrapContent = this.also { wrapContent = value }
 
 /**设置是否缩进内容。*/
-inline infix fun <T> T.indent(value: Boolean) where T : CanIndentContent = this.also { indentContent = value }
+@GeneralDsl
+inline infix fun <T> T.indent(value: Boolean): T where T : CanIndentContent = this.also { indentContent = value }
+
+
+/**换行内容。*/
+@GeneralDsl
+inline fun <T> T.wrap(): T where T : CanWrapContent = this.wrap(true)
+
+/**不换行内容。*/
+@GeneralDsl
+inline fun <T> T.unwrap(): T where T : CanWrapContent = this.wrap(false)
 
 /**缩进内容。*/
-inline fun <T> T.indent() where T : CanIndentContent = this.also { indentContent = true }
+@GeneralDsl
+inline fun <T> T.indent(): T where T : CanIndentContent = this.indent(true)
 
 /**不缩进内容。*/
-inline fun <T> T.unindent() where T : CanIndentContent = this.also { indentContent = false }
+@GeneralDsl
+inline fun <T> T.unindent(): T where T : CanIndentContent = this.indent(false)