+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/dsl/text/MermaidFlowDsl.kt
@@ -1,5 +1,6 @@
 @file:Reference("[Mermaid](https://mermaidjs.github.io)")
 @file:NotImplemented
+@file:Suppress("CanBePrimaryConstructorProperty", "SimpleRedundantLet", "NOTHING_TO_INLINE")
 
 package com.windea.breezeframework.data.dsl.text
 
@@ -9,11 +10,8 @@ import com.windea.breezeframework.data.dsl.*
 import com.windea.breezeframework.data.dsl.text.MermaidFlowConfig.indent
 import com.windea.breezeframework.data.dsl.text.MermaidFlowConfig.quote
 
-//////////Portal function
 
-fun mermaidGraph(direction: MermaidFlowDirection, builder: MermaidFlow.() -> Unit) = MermaidFlow(direction).builder()
-
-///////////////Dsl marker annotations & Dsl element interfaces
+//REGION Dsl marker annotations & Dsl element interfaces
 
 @DslMarker
 annotation class MermaidFlowDsl
@@ -21,31 +19,58 @@ annotation class MermaidFlowDsl
 @MermaidFlowDsl
 interface MermaidFlowDslElement
 
-interface MermaidFlowGraph : CanIndentContent {
-	val nodes: MutableSet<MermaidFlowNode>
-	val links: MutableSet<MermaidFlowLink>
+@MermaidFlowDsl
+abstract class MermaidFlowGraph : CanIndentContent {
+	val nodes: MutableSet<MermaidFlowNode> = mutableSetOf()
+	val links: MutableSet<MermaidFlowLink> = mutableSetOf()
 	
-	fun node() {
-		TODO()
+	
+	@MermaidFlowDsl
+	inline fun node(id: String, text: String? = null, shape: MermaidFlowNodeShape = MermaidFlowNodeShape.Rect) =
+		MermaidFlowNode(id, text, shape).also { nodes += it }
+	
+	@MermaidFlowDsl
+	inline fun link(sourceNodeId: String, targetNodeId: String, linkShape: MermaidFlowLinkShape = MermaidFlowLinkShape.Arrow, linkText: String? = null) =
+		MermaidFlowLink(sourceNodeId, targetNodeId, linkShape, linkText).also { links += it }
+	
+	@MermaidFlowDsl
+	inline fun link(sourceNode: MermaidFlowNode, targetNodeId: String, linkShape: MermaidFlowLinkShape = MermaidFlowLinkShape.Arrow, linkText: String? = null) = run {
+		if(sourceNode !in nodes) nodes += sourceNode
+		MermaidFlowLink(sourceNode.id, targetNodeId, linkShape, linkText).also { links += it }
 	}
 	
-	fun link() {
-		TODO()
+	@MermaidFlowDsl
+	inline fun link(sourceNodeId: String, targetNode: MermaidFlowNode, linkShape: MermaidFlowLinkShape = MermaidFlowLinkShape.Arrow, linkText: String? = null) = run {
+		if(targetNode !in nodes) nodes += targetNode
+		MermaidFlowLink(sourceNodeId, targetNode.id, linkShape, linkText).also { links += it }
+	}
+	
+	@MermaidFlowDsl
+	inline fun link(sourceNode: MermaidFlowNode, targetNode: MermaidFlowNode, linkShape: MermaidFlowLinkShape = MermaidFlowLinkShape.Arrow, linkText: String? = null) = run {
+		if(sourceNode !in nodes) nodes += sourceNode
+		if(targetNode !in nodes) nodes += targetNode
+		MermaidFlowLink(sourceNode.id, targetNode.id, linkShape, linkText).also { links += it }
 	}
 }
 
-////////////Dsl elements & build functions
+//REGION Dsl elements & build functions
 
+@MermaidFlowDsl
+fun mermaidGraph(direction: MermaidFlowDirection, builder: MermaidFlow.() -> Unit) =
+	MermaidFlow(direction).also { it.builder() }
+
+@MermaidFlowDsl
 class MermaidFlow @PublishedApi internal constructor(
-	val direction: MermaidFlowDirection,
-	override val nodes: MutableSet<MermaidFlowNode> = mutableSetOf(),
-	override val links: MutableSet<MermaidFlowLink> = mutableSetOf(),
-	val styles: MutableList<MermaidFlowNodeStyle> = mutableListOf(),
-	val linkStyles: MutableList<MermaidFlowLinkStyle> = mutableListOf(),
-	val classDefs: MutableSet<MermaidFlowClassDef> = mutableSetOf(),
-	val classes: MutableSet<MermaidFlowClass> = mutableSetOf(),
+	direction: MermaidFlowDirection
+) : MermaidFlowGraph(), DslBuilder {
+	val direction: MermaidFlowDirection = direction
+	val styles: MutableList<MermaidFlowNodeStyle> = mutableListOf()
+	val linkStyles: MutableList<MermaidFlowLinkStyle> = mutableListOf()
+	val classDefs: MutableSet<MermaidFlowClassDef> = mutableSetOf()
+	val classRefs: MutableSet<MermaidFlowClassRef> = mutableSetOf()
+	
 	override var indentContent: Boolean = true
-) : MermaidFlowGraph, Dsl {
+	
 	override fun toString(): String {
 		val contentSnippet = buildString {
 			append(nodes.joinToString("\n", "\n\n"))
@@ -53,35 +78,38 @@ class MermaidFlow @PublishedApi internal constructor(
 			append(styles.joinToString("\n", "\n\n"))
 			append(linkStyles.joinToString("\n", "\n\n"))
 			append(classDefs.joinToString("\n", "\n\n"))
-			append(classes.joinToString("\n"))
+			append(classRefs.joinToString("\n"))
 		}
 		val indentedSnippet = if(indentContent) contentSnippet.prependIndent(indent) else contentSnippet
 		return "graph ${direction.text}\n$indentedSnippet"
 	}
 	
-	fun style() {
-		TODO()
-	}
 	
-	fun linkStyle() {
-		TODO()
-	}
+	@MermaidFlowDsl
+	inline fun style(nodeId: String, vararg styles: Pair<String, String>) =
+		MermaidFlowNodeStyle(nodeId, *styles).also { this@MermaidFlow.styles += it }
 	
-	fun classDef() {
-		TODO()
-	}
+	@MermaidFlowDsl
+	inline fun linkStyle(linkOrder: Int, vararg styles: Pair<String, String>) =
+		MermaidFlowLinkStyle(linkOrder, *styles).also { this@MermaidFlow.linkStyles += it }
 	
-	fun classRef() {
-		TODO()
-	}
+	@MermaidFlowDsl
+	inline fun classDef(className: String, vararg styles: Pair<String, String>) =
+		MermaidFlowClassDef(className, *styles).also { this@MermaidFlow.classDefs += it }
+	
+	@MermaidFlowDsl
+	inline fun classRef(vararg nodeIds: String, className: String) =
+		MermaidFlowClassRef(*nodeIds, className = className).also { this@MermaidFlow.classRefs += it }
 }
 
+@MermaidFlowDsl
 class MermaidFlowSubGraph @PublishedApi internal constructor(
-	val name: String,
-	override val nodes: MutableSet<MermaidFlowNode>,
-	override val links: MutableSet<MermaidFlowLink>,
+	name: String
+) : MermaidFlowGraph() {
+	val name: String = name //NOTE do not ensure that argument is valid
+	
 	override var indentContent: Boolean = true
-) : MermaidFlowGraph {
+	
 	override fun toString(): String {
 		val contentSnippet = buildString {
 			append(nodes.joinToString("\n", "\n\n"))
@@ -92,74 +120,100 @@ class MermaidFlowSubGraph @PublishedApi internal constructor(
 	}
 }
 
+@MermaidFlowDsl
 class MermaidFlowNode @PublishedApi internal constructor(
-	val id: String,
-	val text: String?,
-	val shape: MermaidFlowNodeShape = MermaidFlowNodeShape.Rect
+	id: String,
+	text: String? = null,
+	shape: MermaidFlowNodeShape = MermaidFlowNodeShape.Rect
 ) : MermaidFlowDslElement {
+	val id: String = id //NOTE do not ensure that argument is valid
+	val text: String? = text //NOTE do not ensure that argument is valid
+	val shape: MermaidFlowNodeShape = shape
+	
 	//TODO omit text, no double quote surround text if not necessary
 	override fun toString(): String {
-		val textSnippet = (text ?: id).let { "$quote$it$quote" }
+		val textSnippet = (text ?: id).wrapQuote(quote)
 		return "$id${shape.prefix}$textSnippet${shape.suffix}"
 	}
 }
 
+@MermaidFlowDsl
 class MermaidFlowLink @PublishedApi internal constructor(
-	val sourceNode: MermaidFlowNode,
-	val targetNode: MermaidFlowNode,
-	val linkText: String?,
-	val linkShape: MermaidFlowLinkShape
+	sourceNodeId: String,
+	targetNodeId: String,
+	linkShape: MermaidFlowLinkShape = MermaidFlowLinkShape.Arrow,
+	linkText: String? = null
 ) : MermaidFlowDslElement {
+	val sourceNodeId: String = sourceNodeId //NOTE do not ensure that argument is valid
+	val targetNodeId: String = targetNodeId //NOTE do not ensure that argument is valid
+	val linkShape: MermaidFlowLinkShape = linkShape
+	val linkText: String? = linkText //NOTE do not ensure that argument is valid
+	
 	//TODO escape if necessary
 	override fun toString(): String {
-		val sourceSnippet = sourceNode.text?.let { sourceNode.toString() } ?: sourceNode.id
-		val targetSnippet = targetNode.text?.let { targetNode.toString() } ?: targetNode.id
-		val linkSnippet = linkText?.let { "${linkShape.prefix} $linkText ${linkShape.suffix}" } ?: linkShape.delimiter
-		return "$sourceSnippet $linkSnippet $targetSnippet"
+		val linkSnippet = linkShape.delimiter
+		val linkTextSnippet = linkText?.let { "|$linkText|" } ?: ""
+		return "$sourceNodeId $linkSnippet$linkTextSnippet $targetNodeId"
 	}
 }
 
+@MermaidFlowDsl
 class MermaidFlowNodeStyle @PublishedApi internal constructor(
-	val nodeId: String,
-	val styleMap: Map<String, String> = mapOf()
+	nodeId: String,
+	vararg styles: Pair<String, String>
 ) : MermaidFlowDslElement {
+	val nodeId: String = nodeId //NOTE do not ensure that argument is valid
+	val styles: Map<String, String> = styles.toMap() //NOTE do not ensure that argument is valid
+	
 	override fun toString(): String {
-		val styleMapSnippet = styleMap.joinToString { (k, v) -> "$k: $v" }
+		val styleMapSnippet = styles.joinToString { (k, v) -> "$k: $v" }
 		return "style $nodeId $styleMapSnippet"
 	}
 }
 
+@MermaidFlowDsl
 class MermaidFlowLinkStyle @PublishedApi internal constructor(
-	val linkOrder: Int,
-	val styleMap: Map<String, String> = mapOf()
+	linkOrder: Int,
+	vararg styles: Pair<String, String>
 ) : MermaidFlowDslElement {
+	val linkOrder: Int = linkOrder //NOTE do not ensure that argument is valid
+	val styles: Map<String, String> = styles.toMap() //NOTE do not ensure that argument is valid
+	
 	override fun toString(): String {
-		val styleMapSnippet = styleMap.joinToString { (k, v) -> "$k: $v" }
+		val styleMapSnippet = styles.joinToString { (k, v) -> "$k: $v" }
 		return "style $linkOrder $styleMapSnippet"
 	}
 }
 
+@MermaidFlowDsl
 class MermaidFlowClassDef @PublishedApi internal constructor(
-	val className: String,
-	val styleMap: Map<String, String> = mapOf()
+	className: String,
+	vararg styles: Pair<String, String>
 ) : MermaidFlowDslElement {
+	val className: String = className //NOTE do not ensure that argument is valid
+	val styles: Map<String, String> = styles.toMap() //NOTE do not ensure that argument is valid
+	
 	override fun toString(): String {
-		val styleMapSnippet = styleMap.joinToString { (k, v) -> "$k: $v" }
+		val styleMapSnippet = styles.joinToString { (k, v) -> "$k: $v" }
 		return "classDef $className $styleMapSnippet"
 	}
 }
 
-class MermaidFlowClass @PublishedApi internal constructor(
-	val nodeIds: Set<String>,
-	val className: String
+@MermaidFlowDsl
+class MermaidFlowClassRef @PublishedApi internal constructor(
+	vararg nodeIds: String,
+	className: String
 ) : MermaidFlowDslElement {
+	val nodeIds = nodeIds.toSet() //NOTE do not ensure that argument is valid
+	val className = className //NOTE do not ensure that argument is valid
+	
 	override fun toString(): String {
-		val nodeIdsSnippet = nodeIds.joinToString()
-		return "class $nodeIdsSnippet $className"
+		val nodeIdSetSnippet = nodeIds.joinToString()
+		return "class $nodeIdSetSnippet $className"
 	}
 }
 
-////////enumerations and constants
+//REGION Enumerations and constants
 
 enum class MermaidFlowDirection(
 	val text: String
@@ -184,20 +238,15 @@ enum class MermaidFlowNodeShape(
 }
 
 enum class MermaidFlowLinkShape(
-	val suffix: String,
-	val prefix: String,
 	val delimiter: String
 ) {
-	Normal("---", "--", "---"),
-	Arrow("-->", "--", "-->"),
-	Dotted("-.->", "-.", ".->"),
-	Thick("==>", "==", "==>")
+	Arrow("-->"),
+	Dotted("-.->"),
+	Thick("==>"),
+	Normal("---")
 }
 
-//////////TODO Param handler extensions
-
-
-//////////Config object
+//REGION Config object
 
 object MermaidFlowConfig : DslConfig {
 	/**缩进长度。*/
@@ -208,6 +257,4 @@ object MermaidFlowConfig : DslConfig {
 	
 	internal val indent get() = if(indentSize <= -1) "\t" * indentSize else " " * indentSize
 	internal val quote get() = if(useDoubleQuote) "\"" else "'"
-	
-	inline operator fun invoke(builder: MermaidFlowConfig.() -> Unit) = this.builder()
 }