+++ breeze-javafx/src/main/kotlin/com/windea/breezeframework/javafx/extensions/TupleExtensions.kt
@@ -1,0 +1,5 @@
+package com.windea.breezeframework.javafx.extensions
+
+import javafx.geometry.*
+
+fun Pair<Double, Double>.toPoint2D() = Point2D(this.first, this.second)