+++ CHANGELOG.md
@@ -85,6 +85,8 @@
 * [X] 更新版本号。
 * [X] breeze-core 添加`Collection.randomOrNull()`扩展。
 * [X] breeze-core 重构系统扩展。
+* [X] breeze-core 重构与重新格式化代码。
+* [X] breeze-data 重构XmlDsl和MermaidFlowDsl。
 
 * [ ] breeze-data 提供mermaid和plantUml的Dsl。
     * [ ] UPDATING 提供mermaid flow chart的Dsl。